backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-dotenv==0.19.0
openai==0.27.0
PyPDF2==3.0.1
pydantic==1.8.2 